src/kudu/benchmarks/ycsb/requirements.txt
matplotlib
numpy
pyYAML